requirements.txt
uvicorn==0.30.1
gunicorn==22.0.0

fastapi==0.115.6
pydantic==2.10.4

python-dotenv==1.0.1
email-validator==2.2.0
python-multipart==0.0.20
argon2-cffi==23.1.0